Update dockyards cluster controller to reconcile ingress-nginx ip
Currently changes to the gateway addresses in the management cluster
will not trigger any updates to the services in the workload clusters.
This causes issues with dns records not being updated to the new gateway
addresses.

This updates the dockyards cluster controller to reconcile the
ingress-nginx service to specify the load balancer ip. This does not
solve the issue directly but will trigger workload service
reconciliation in the management cluster.