Update dockyards node pool controller to reconcile tls route condition
Currently the talos control plane has a requirement on an existing tls
route to reconcile. If the tls route does not exist the talos control
plane reconciliation is ignored but this places the dockyards node pool
in a ready state.

This adds a negative condition on the dockyards node pool when the tls
route can not be found.