Update dockyards cluster controller to use gateway for api endpoint
Currently the dockyards cluster controller uses a service resource to
lookup the external name to use for cluster api endpoints. This service
resource is redundant as the information is already available in the
gateway object that the cluster will attach to.

This updates the dockyards cluster controller to use the gateway
resource to lookup the external name to use for api endpoints.