Update dockyards cluster controller to use custom name
Currently the dockyards cluster controller get a default name based on
the kind. This causes issues with the updated controller-runtime
dependency since both the cluster-api and dockyards cluster controllers
gets the same name.

This adds an explicit name to the dockyards cluster controller to fix
the name collision with the cluster-api controller.