Update dockyards release controller to reconcile data source
Currently the dockyards release controller will try to update the data
volume spec when the parent dockyards release updates. This will fail
because the data volume spec is immutable.

This updates the dockyards release controller to reconcile a unique data
volume for each dockyards release version. This adds a data source
resource with the same name as the dockyards release to use for cloning.