Update dockyards workload controller to filter address families
Currently the workloads clusters are single stacked to IPv4. This causes
issue when the management cluster has IPv6 enabled and chooses the first
available address on gateway resources.

This updates the dockyards workload controller to filter any gateway
addresses that are not IPv4. This is a temporary measure until workload
clusters get dual stack support.