Update dockyards node pool controller to support cluster subnets
Currently the dockyards node pool controller uses a hardcoded list of
subnets to use for talos configs.

This adds support for the pod and service subnet fields on the dockyards
cluster resources to allow specifying desired configuration.

As a side effect the default behaviour of a dockyards cluster without
any subnets specified will be to use the default talos subnets.