Update dockyards node pool controller to support multus as default
Currently the dockyards node pool controller will always add the pod
network to kubevirt machine templates when multus support is enabled.

This adds support for a custom label on network attachment definitions
to add the multus network as default. If any network attachment
definitions in an organization has the
`kubevirt.dockyards.io/network-as-default` label no pod network will be
added to the reconciled kubevirt machine template.