Update dockyards node pool controller to use namespaced network names
Currently the multus network names are referenced without namespace to
implicitly refer to the same namespace as the reconciled kubevirt
machine templates. This causes issues together with default multus
networks and errors around network attachment definitions not found in
the kube-system namespace.

This updates the dockyards node pool to include namespace in network
names to ensure that the network attachment definition is explicit.