Update valid node ip subnet path in config patches
Currently the dockyards node pool controller generates config patches
for the valid node ip subnets that references a nested path that does
not exist in the default talos config. This causes the generation of
bootstrap configuration to fail.

This updates the path of the config patch to reference a path known to
exist.