Update dockyards cluster controller to optionally reconcile ingress
Currently the dockyards cluster controller will always reconcile the
ingress workload to use the gateway address in the workload cluster.
This causes issues when the workload clusters should have dedicated load
balancer configuration.

This updates the dockyards cluster controller to only reconcile the load
balancer ip of the ingress workload when enabled.

This also updates the dockyards cluster controller to add a condition
for the ingress workload reconciliation to simplify troubleshooting.