Update dockyards node pool to mark conditions on talos control plane
Currently there are several cases that could throw errors in the
dockyards node pool controller related to the talos control plane. These
cases are only logged and not marked on the dockyards node pool resource
itself.

This updates the behaviour to mark all errors related to reconciling the
talos control plane.